Separated ROI selection mesh modifier from widget
