+++ snp_tpp/CMakeLists.txt
@@ -50,10 +50,13 @@ ament_target_dependencies(
 
 # Custom mesh modifier
 qt5_wrap_ui(${PROJECT_NAME}_ui_mocs src/roi_selection_mesh_modifier_widget.ui)
-qt5_wrap_cpp(${PROJECT_NAME}_mocs src/roi_selection_mesh_modifier.h)
-
-add_library(${PROJECT_NAME}_mesh_modifier SHARED src/roi_selection_mesh_modifier.cpp ${${PROJECT_NAME}_ui_mocs}
-                                                 ${${PROJECT_NAME}_mocs})
+qt5_wrap_cpp(${PROJECT_NAME}_mocs src/roi_selection_mesh_modifier_widget.h)
+add_library(
+  ${PROJECT_NAME}_mesh_modifier SHARED
+  src/roi_selection_mesh_modifier.cpp
+  src/roi_selection_mesh_modifier_widget.cpp
+  ${${PROJECT_NAME}_ui_mocs}
+  ${${PROJECT_NAME}_mocs})
 target_include_directories(
   ${PROJECT_NAME}_mesh_modifier PUBLIC "$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>"
                                        "$<BUILD_INTERFACE:${CMAKE_CURRENT_BINARY_DIR}>" "$<INSTALL_INTERFACE:include>")
